Splitting out waitContainerToCatchUp into its own file/module
